Z_CheckBox/requirements.txt
python-dotenv==1.0.1
requests==2.32.3
mysql-connector-python==9.0.0
